Move subprocess helper functions to theano.utils
These helper functions where in `theano.misc.windows`, but they're not actually
specific to Windows and they're clearly generic "utils" style functions.
